requirements.txt
pillow~=11.3.0
python-barcode~=0.15.1
python-telegram-bot==20.3
